Refactored pubsub module
No changes to module logic was added. The module was just split up for
better maintainability
